Created an angular 2 item count control
