+++ bootstrapper/cards/cardsNg2.html
@@ -51,5 +51,6 @@
 		<rlPager></rlPager>
 		<rlPageSize></rlPageSize>
 		<rlSelectionControl></rlSelectionControl>
+		<rlItemCount></rlItemCount>
 	</rlCardContainer>
 </div>
