+++ bootstrapper/cards/cardsNg2Bootstrapper.ts
@@ -18,6 +18,7 @@ import {
 	FILTER_DIRECTIVES,
 } from '../../source/components/cardContainer/filters/index';
 import { CardContainerComponent } from '../../source/components/cardContainer/cardContainer';
+import { ItemCountComponent } from '../../source/components/cardContainer/itemCount/itemCount';
 import { PagerComponent } from '../../source/components/cardContainer/pager/pager';
 import { PageSizeComponent } from '../../source/components/cardContainer/pageSize/pageSize';
 import { SelectionComponent } from '../../source/components/cardContainer/selectionControl/selectionControl';
@@ -30,6 +31,7 @@ import { SelectionComponent } from '../../source/components/cardContainer/select
 		INPUT_DIRECTIVES,
 		FILTER_DIRECTIVES,
 		CardContainerComponent,
+		ItemCountComponent,
 		PagerComponent,
 		PageSizeComponent,
 		SelectionComponent,
