+++ source/components/cardContainer/itemCount/itemCount.html
@@ -1,3 +1,3 @@
-<p ng-show="!itemCount.cardContainer.dataSource.loadingDataSet">
-	Showing <strong>{{itemCount.cardContainer.dataSource.dataSet.length}} of {{itemCount.cardContainer.dataSource.count}}</strong> total items
+<p *ngIf="!cardContainer?.dataSource?.loadingDataSet">
+	Showing <strong>{{cardContainer?.dataSource?.dataSet?.length}} of {{itemCount.cardContainer.dataSource.count}}</strong> total items
 </p>
