+++ source/components/cardContainer/itemCount/itemCount.ng1.html
@@ -1,0 +1,3 @@
+<p ng-show="!itemCount.cardContainer.dataSource.loadingDataSet">
+	Showing <strong>{{itemCount.cardContainer.dataSource.dataSet.length}} of {{itemCount.cardContainer.dataSource.count}}</strong> total items
+</p>
